+++ packages/atomic/src/components/commerce/index.ts
@@ -2,6 +2,7 @@
 export {AtomicCommerceFacet} from './atomic-commerce-facet/atomic-commerce-facet.js';
 export {AtomicCommerceInterface} from './atomic-commerce-interface/atomic-commerce-interface.js';
 export {AtomicCommercePager} from './atomic-commerce-pager/atomic-commerce-pager.js';
+export {AtomicCommerceProductList} from './atomic-commerce-product-list/atomic-commerce-product-list.js';
 export {AtomicCommerceSearchBoxInstantProducts} from './atomic-commerce-search-box-instant-products/atomic-commerce-search-box-instant-products.js';
 export {AtomicCommerceSearchBoxQuerySuggestions} from './atomic-commerce-search-box-query-suggestions/atomic-commerce-search-box-query-suggestions.js';
 export {AtomicCommerceSearchBoxRecentQueries} from './atomic-commerce-search-box-recent-queries/atomic-commerce-search-box-recent-queries.js';