+++ packages/quantic/force-app/main/default/lwc/quanticResultList/e2e/quanticResultList.e2e.ts
@@ -38,7 +38,7 @@ useCaseTestCases.forEach((useCase) => {
             fieldsToInclude: expect.arrayContaining(expectedFieldsToInclude),
           })
         );
-        expect(resultList.resultList).toBeVisible();
+        await expect(resultList.resultList).toBeVisible();
       });
     });